requirements.txt
flask-bootstrap
#bootstrap-flask
flask-sqlalchemy
flask-api
redis
celery
#gpiozero
requests
rpi.gpio
pi-plates
#spidev
#six
